NJU-ProjectN/nanos-lite ics2021 initialized
NJU-ProjectN/nanos-lite 2a141760e31be246a7316942293a97873925bc2f Makefile: use header files in newlib-navy

+++ abstract-machine/am/src/riscv/riscv.h
@@ -12,7 +12,7 @@ static inline void outb(uintptr_t addr, uint8_t  data) { *(volatile uint8_t  *)a
 static inline void outw(uintptr_t addr, uint16_t data) { *(volatile uint16_t *)addr = data; }
 static inline void outl(uintptr_t addr, uint32_t data) { *(volatile uint32_t *)addr = data; }
 
-static inline uint32_t ind(uintptr_t addr) { return *(volatile uint64_t *)addr; }
+static inline uint64_t ind(uintptr_t addr) { return *(volatile uint64_t *)addr; }
 static inline void outd(uintptr_t addr, uint64_t data) { *(volatile uint64_t *)addr = data; }
 
 #define PTE_V 0x01
